+++ solidity/contracts/interfaces/IBancorFormula.sol
@@ -4,6 +4,6 @@ pragma solidity ^0.4.11;
     Bancor Formula interface
 */
 contract IBancorFormula {
-    function calculatePurchaseReturn(uint256 _supply, uint256 _reserveBalance, uint256 _reserveRatio, uint256 _depositAmount) public constant returns (uint256);
-    function calculateSaleReturn(uint256 _supply, uint256 _reserveBalance, uint256 _reserveRatio, uint256 _sellAmount) public constant returns (uint256);
+    function calculatePurchaseReturn(uint256 _supply, uint256 _reserveBalance, uint32 _reserveRatio, uint256 _depositAmount) public constant returns (uint256);
+    function calculateSaleReturn(uint256 _supply, uint256 _reserveBalance, uint32 _reserveRatio, uint256 _sellAmount) public constant returns (uint256);
 }
